Data creation now working.
