--- src/server/db/fakeDataGenerator.js
@@ -1,26 +1,0 @@
-/* NPM Installation Dependencies */
-const faker = require('faker');
-
-const generateFakeData = () => {
-  generateFakeUsers();
-  generateFakeOrganizations();
-  generateFakeEvents();
-}
-
-const gradYears = [2019, 2020, 2021, 2022];
-
-const generateFakeUsers = () => {
-
-}
-
-const generateFakeOrganizations = () => {
-
-}
-
-const generateFakeEvents = () => {
-
-}
-
-module.exports = {
-  generateFakeData
-}
